Add span tracing logic here
If this functions handles all the API http responses,
it'll be convenient to add the tracing output here,
wether we respond with an error or a success response,
both responses are properly traced depending on its type.

When an error occurs, we tell the span that there was an error.
On each response we tell the span the returned http status code.